backend/requirements.txt
# Azure Functions runtime
azure-functions>=1.17.0

# Azure OpenAI
openai>=1.12.0

# Azure Blob Storage
azure-storage-blob>=12.19.0

# RAG pipeline (Phase 2)
faiss-cpu>=1.7.4
sentence-transformers>=2.5.0,<3.0.0
transformers>=4.38.0,<5.0.0

# Document extraction (Phase 2)
PyMuPDF>=1.23.0
python-docx>=1.1.0

# Data validation
pydantic>=2.6.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
